mlproject/requirement.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-cronjobs==0.2.3
django-crontab==0.7.1
djangorestframework==3.15.1
idna==3.7
numpy==1.24.4
panda==0.3.1
requests==2.31.0
sqlparse==0.5.0
typing-extensions==4.11.0
urllib3==2.2.1
